Summarize of controller and observer implementations in one file and inluding in one python module
